+++ src/main/c/jni/org_jpy_PyLib.c
@@ -499,12 +499,20 @@ void dumpDict(const char* dictName, PyObject* dict)
 
     size = PyDict_Size(dict);
     printf(">> dumpDict: %s.size = %ld\n", dictName, size);
+#if PY_VERSION_HEX >= 0x030D0000 // >=3.13
+    // PyDict_Next is not thread-safe, so we need to protect it with a critical section
+    // https://docs.python.org/3/howto/free-threading-extensions.html#pydict-next
+    Py_BEGIN_CRITICAL_SECTION(dict);
+#endif
     while (PyDict_Next(dict, &pos, &key, &value)) {
         const char* name;
         name = JPy_AS_UTF8(key);
         printf(">> dumpDict: %s[%ld].name = '%s'\n", dictName, i, name);
         i++;
     }
+#if PY_VERSION_HEX >= 0x030D0000 // >=3.13
+    Py_END_CRITICAL_SECTION();
+#endif
 }
 
 /**
@@ -521,7 +529,7 @@ PyObject *getMainGlobals() {
     }
 
     pyGlobals = PyModule_GetDict(pyMainModule); // borrowed ref
-    JPy_INCREF(pyGlobals);
+    JPy_XINCREF(pyGlobals);
 
     return pyGlobals;
 }
@@ -557,7 +565,7 @@ JNIEXPORT jobject JNICALL Java_org_jpy_PyLib_getCurrentGlobals
 
     JPy_BEGIN_GIL_STATE
 
-#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION <= 12
+#if PY_VERSION_HEX < 0x030D0000 // < 3.13
     globals = PyEval_GetGlobals(); // borrowed ref
     JPy_XINCREF(globals);
 #else
@@ -588,7 +596,7 @@ JNIEXPORT jobject JNICALL Java_org_jpy_PyLib_getCurrentLocals
 
     JPy_BEGIN_GIL_STATE
 
-#if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION <= 12
+#if PY_VERSION_HEX < 0x030D0000 // < 3.13
     locals = PyEval_GetLocals(); // borrowed ref
     JPy_XINCREF(locals);
 #else
@@ -1124,7 +1132,7 @@ JNIEXPORT void JNICALL Java_org_jpy_PyLib_incRef
     if (Py_IsInitialized()) {
         JPy_BEGIN_GIL_STATE
 
-        refCount = pyObject->ob_refcnt;
+        refCount = Py_REFCNT(pyObject);
         JPy_DIAG_PRINT(JPy_DIAG_F_MEM, "Java_org_jpy_PyLib_incRef: pyObject=%p, refCount=%d, type='%s'\n", pyObject, refCount, Py_TYPE(pyObject)->tp_name);
         JPy_INCREF(pyObject);
 
@@ -1150,7 +1158,7 @@ JNIEXPORT void JNICALL Java_org_jpy_PyLib_decRef
     if (Py_IsInitialized()) {
         JPy_BEGIN_GIL_STATE
 
-        refCount = pyObject->ob_refcnt;
+        refCount = Py_REFCNT(pyObject);
         if (refCount <= 0) {
             JPy_DIAG_PRINT(JPy_DIAG_F_ALL, "Java_org_jpy_PyLib_decRef: error: refCount <= 0: pyObject=%p, refCount=%d\n", pyObject, refCount);
         } else {
@@ -1183,7 +1191,7 @@ JNIEXPORT void JNICALL Java_org_jpy_PyLib_decRefs
         buf = (*jenv)->GetLongArrayElements(jenv, objIds, &isCopy);
         for (i = 0; i < len; i++) {
             pyObject = (PyObject*) buf[i];
-            refCount = pyObject->ob_refcnt;
+            refCount = Py_REFCNT(pyObject);
             if (refCount <= 0) {
                 JPy_DIAG_PRINT(JPy_DIAG_F_ALL, "Java_org_jpy_PyLib_decRefs: error: refCount <= 0: pyObject=%p, refCount=%d\n", pyObject, refCount);
             } else {